Moved minigame model to its own folder
