+++ Protocols.External/[e] EmailHandler/[d] DeleteMessage.cs
@@ -1,4 +1,6 @@
-﻿using Server.Reawakened.Network.Extensions;
+﻿using A2m.Server.MessageCenter.Proto;
+using A2m.Server;
+using Server.Reawakened.Network.Extensions;
 using Server.Reawakened.Network.Protocols;
 using Server.Reawakened.Players.Extensions;
 using Server.Reawakened.XMLs.Bundles;
@@ -19,18 +21,23 @@ public class DeleteMessage : ExternalProtocol
         {
             var item = ItemCatalog.GetItemFromId(Player.Character.EmailMessages[messageId].Item.ItemId);
 
-            Task.Delay(3300); //Adds item after open gift animation.
-
-            Player.Character.AddItem(item, item.ItemNumber);
-            Player.SendUpdatedInventory(false);
-
-            var mailMessage = Player.Character.EmailMessages[messageId];
-            var mail = Player.Character.Emails[messageId];
-
-            Player.Character.EmailMessages.Remove(mailMessage);
-            Player.Character.Emails.Remove(mail);
-
-            Player.SendXt("ed", messageId);
+            _ = RunGiftAnimation(item, messageId);
         }
+    }
+
+    public async Task RunGiftAnimation(ItemDescription item, int messageId)
+    {
+        await Task.Delay(3300); //Adds item after open gift animation.
+
+        Player.Character.AddItem(item, item.ItemNumber);
+        Player.SendUpdatedInventory(false);
+
+        var mailMessage = Player.Character.EmailMessages[messageId];
+        var mail = Player.Character.Emails[messageId];
+
+        Player.Character.EmailMessages.Remove(mailMessage);
+        Player.Character.Emails.Remove(mail);
+
+        Player.SendXt("ed", messageId);
     }
 }
