+++ Server.Reawakened/Players/Models/Minigames/MinigameModel.cs
@@ -1,7 +1,8 @@
 ﻿using Server.Reawakened.Players.Helpers;
+using Server.Reawakened.Players.Models.Character;
 using Server.Reawakened.XMLs.BundlesInternal;
 
-namespace Server.Reawakened.Players.Models.Character;
+namespace Server.Reawakened.Players.Models.Minigames;
 public class ArenaModel
 {
     public bool ShouldStartArena { get; set; }
@@ -15,6 +16,7 @@ public class ArenaModel
     public static void SetCharacterIds(Player player, IEnumerable<Player> players)
     {
         var playersInGroup = players.ToArray();
+
         player.TempData.ArenaModel.FirstPlayerId = playersInGroup.Length > 0 ? playersInGroup[0].GameObjectId : 0;
         player.TempData.ArenaModel.SecondPlayerId = playersInGroup.Length > 1 ? playersInGroup[1].GameObjectId : 0;
         player.TempData.ArenaModel.ThirdPlayerId = playersInGroup.Length > 2 ? playersInGroup[2].GameObjectId : 0;
@@ -31,7 +33,6 @@ public class ArenaModel
                 itemsGotten.Add(item);
 
         var randomItemReward = itemsGotten[random.Next(itemsGotten.Count)].ItemId;
-
         var itemsLooted = FormatItemString(randomItemReward, 1);
 
         return itemsLooted.ToString();
@@ -42,11 +43,9 @@ public class ArenaModel
         var lootableItems = new SeparatedStringBuilder('|');
 
         if (LootCatalog.LootCatalog.TryGetValue(arenaId, out var value))
-        {
             foreach (var reward in value.ItemRewards)
                 foreach (var itemReward in reward.Items)
                     lootableItems.Append(itemReward.ItemId);
-        }
 
         return lootableItems.ToString();
     }
